Move hooks to helper
